src/app 폴더에 기본 index.tsx 생성 후 react 기본 render에서 호출
