+++ src/app/index.tsx
@@ -1,0 +1,23 @@
+import { BrowserRouter as Router } from "react-router-dom"
+import Header from "../widgets/ui/Header.tsx"
+import Footer from "../widgets/ui/Footer.tsx"
+import PostsManagerPage from "../pages/PostsManagerPage.tsx"
+import React from "react"
+
+const App = () => {
+  return (
+    <React.StrictMode>
+      <Router>
+        <div className="flex flex-col min-h-screen">
+          <Header />
+          <main className="flex-grow container mx-auto px-4 py-8">
+            <PostsManagerPage />
+          </main>
+          <Footer />
+        </div>
+      </Router>
+    </React.StrictMode>
+  )
+}
+
+export default App;
